@@ -1,14 +1,19 @@
+import 'dart:developer';
+
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:patient_management_app/blocs/base_state.dart';
 import 'package:patient_management_app/blocs/patient/patient_event.dart';
 import 'package:patient_management_app/blocs/patient/patient_state.dart';
-import 'package:patient_management_app/blocs/base_state.dart';
+import 'package:patient_management_app/config/constants.dart';
 import 'package:patient_management_app/models/patient.dart';
 import 'package:patient_management_app/services/patient_service.dart';
+import 'package:patient_management_app/utils/cache_manager.dart';
 
 class PatientBloc extends Bloc<PatientEvent, PatientState> {
   final PatientService _patientService;
+  final CacheManager cacheManager = CacheManager();
 
-  PatientBloc({PatientService? patientService}) 
+  PatientBloc({PatientService? patientService})
       : _patientService = patientService ?? PatientService(),
         super(const PatientState()) {
     on<PatientFetchAll>(_onFetchAll);
@@ -21,9 +26,22 @@ class PatientBloc extends Bloc<PatientEvent, PatientState> {
 
   Future<void> _onFetchAll(PatientFetchAll event, Emitter<PatientState> emit) async {
     emit(state.copyWith(status: Status.loading));
-    
+
+    final cachedPatients = cacheManager.getData(key: CacheKeys.patients);
+
+    if (cachedPatients != null && cachedPatients is List<Patient> && cachedPatients.isNotEmpty) {
+      emit(state.copyWith(
+        status: Status.success,
+        patients: cachedPatients,
+        clearFailure: true,
+      ));
+
+      log("Patients were loaded from cache!");
+    }
+
     try {
       final patients = await _patientService.getAllPatients();
+      cacheManager.cacheData(key: CacheKeys.patients, data: patients);
       emit(state.copyWith(
         status: Status.success,
         patients: patients,
@@ -39,7 +57,7 @@ class PatientBloc extends Bloc<PatientEvent, PatientState> {
 
   Future<void> _onFetchOne(PatientFetchOne event, Emitter<PatientState> emit) async {
     emit(state.copyWith(status: Status.loading));
-    
+
     try {
       final patient = await _patientService.getPatient(event.patientId);
       emit(state.copyWith(
@@ -60,12 +78,12 @@ class PatientBloc extends Bloc<PatientEvent, PatientState> {
       status: Status.loading,
       searchQuery: event.query,
     ));
-    
+
     try {
-      final patients = await _patientService.searchPatients(event.query);
+      // final patients = await _patientService.searchPatients(event.query);
       emit(state.copyWith(
         status: Status.success,
-        patients: patients,
+        patients: state.patients.where((element) => element.fullName.contains(event.query)).toList(),
         clearFailure: true,
       ));
     } catch (e) {
@@ -78,10 +96,10 @@ class PatientBloc extends Bloc<PatientEvent, PatientState> {
 
   Future<void> _onCreate(PatientCreate event, Emitter<PatientState> emit) async {
     emit(state.copyWith(status: Status.loading));
-    
+
     try {
       // Convert map to Patient object
-      final now = DateTime.now().toIso8601String();
+      final now = DateTime.now();
       final patient = Patient(
         fullName: event.patientData['fullName'],
         age: event.patientData['age'],
@@ -90,15 +108,15 @@ class PatientBloc extends Bloc<PatientEvent, PatientState> {
         mobileNumber: event.patientData['mobileNumber'],
         pastIllnesses: event.patientData['pastIllnesses'] ?? '',
         status: event.patientData['status'] ?? 'active',
-        createdAt: now,
-        updatedAt: now,
+        // createdAt: now,
+        // updatedAt: now,
       );
-      
+
       final createdPatient = await _patientService.createPatient(patient);
-      
+
       // Update the list with the new patient
       final updatedPatients = List<Patient>.from(state.patients)..add(createdPatient);
-      
+
       emit(state.copyWith(
         status: Status.success,
         patients: updatedPatients,
@@ -115,7 +133,7 @@ class PatientBloc extends Bloc<PatientEvent, PatientState> {
 
   Future<void> _onUpdate(PatientUpdate event, Emitter<PatientState> emit) async {
     emit(state.copyWith(status: Status.loading));
-    
+
     try {
       // Get current patient
       Patient? currentPatient;
@@ -124,13 +142,13 @@ class PatientBloc extends Bloc<PatientEvent, PatientState> {
       } else {
         currentPatient = await _patientService.getPatient(event.patientId);
       }
-      
+
       if (currentPatient == null) {
         throw Exception('Patient not found');
       }
-      
+
       // Create updated patient
-      final now = DateTime.now().toIso8601String();
+      final now = DateTime.now();
       final updatedPatient = Patient(
         id: currentPatient.id,
         fullName: event.patientData['fullName'] ?? currentPatient.fullName,
@@ -140,19 +158,19 @@ class PatientBloc extends Bloc<PatientEvent, PatientState> {
         mobileNumber: event.patientData['mobileNumber'] ?? currentPatient.mobileNumber,
         pastIllnesses: event.patientData['pastIllnesses'] ?? currentPatient.pastIllnesses,
         status: event.patientData['status'] ?? currentPatient.status,
-        createdAt: currentPatient.createdAt,
-        updatedAt: now,
+        // createdAt: currentPatient.createdAt,
+        // updatedAt: now,
         recordsCount: currentPatient.recordsCount,
         lastVisit: currentPatient.lastVisit,
       );
-      
+
       final savedPatient = await _patientService.updatePatient(updatedPatient);
-      
+
       // Update the list with the updated patient
       final updatedPatients = state.patients.map((p) {
         return p.id == savedPatient.id ? savedPatient : p;
       }).toList();
-      
+
       emit(state.copyWith(
         status: Status.success,
         patients: updatedPatients,
@@ -169,13 +187,13 @@ class PatientBloc extends Bloc<PatientEvent, PatientState> {
 
   Future<void> _onDelete(PatientDelete event, Emitter<PatientState> emit) async {
     emit(state.copyWith(status: Status.loading));
-    
+
     try {
       await _patientService.deletePatient(event.patientId);
-      
+
       // Remove the deleted patient from the list
       final updatedPatients = state.patients.where((p) => p.id != event.patientId).toList();
-      
+
       emit(state.copyWith(
         status: Status.success,
         patients: updatedPatients,
